pg_journal: add extra data to the message (errcode and cursor position)
This nominally breaks the idea of pg_journal, but causes the journal
output to more closely match the normal server logging (thus improving
usability).

I plan to add a catalog file making use of extra fields when they exist,
but the SQL error code and the cursor position are not worth additional
line breaks.